Change "revenue" property of TmdbMovie from int to Int64 to deal with larger numbers (thanks Venom_LoD)

+++ WatTmdb/trunk/WatTmdb/V3/TmdbMovie.cs
@@ -54,7 +54,7 @@ namespace WatTmdb.V3
         public List<ProductionCompany> production_companies { get; set; }
         public List<ProductionCountry> production_countries { get; set; }
         public string release_date { get; set; }
-        public int revenue { get; set; }
+        public Int64 revenue { get; set; }
         public int runtime { get; set; }
         public List<SpokenLanguage> spoken_languages { get; set; }
         public string tagline { get; set; }
